Phuong | Finish setting d3.js library
Create the test chart for testing d3.js
